server/pyproject.toml
[project]
name = "oracy-server"
version = "0.1.0"
description = "Oracy AI Backend - AI-powered oracy assessment platform"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Oracy AI Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "Programming Language :: Python :: 3.12",
    "Framework :: Django :: 5.0",
]

# Core dependencies
dependencies = [
    # Django & DRF
    "django>=5.0,<5.1",
    "djangorestframework>=3.15.0",
    "django-cors-headers>=4.3.0",
    "django-filter>=24.0",
    "django-guardian>=2.4.0",
    "django-ratelimit>=4.1.0",
    
    # Database
    "psycopg[binary]>=3.1.0",
    
    # Caching & Message Queue
    "redis>=5.0.0",
    "django-redis>=5.4.0",
    
    # Celery
    "celery>=5.3.0",
    "django-celery-beat>=2.6.0",
    "django-celery-results>=2.5.0",
    
    # Authentication
    "PyJWT>=2.8.0",
    "argon2-cffi>=23.1.0",
    
    # API Documentation
    "drf-spectacular>=0.27.0",
    
    # Storage
    "django-storages>=1.14.0",
    "boto3>=1.34.0",
    "botocore>=1.34.0",
    
    # AI/ML Services
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "litellm>=1.0.0",
    "instructor>=0.4.0",
    
    # Audio Processing
    "librosa>=0.10.0",
    "pydub>=0.25.0",
    "soundfile>=0.12.0",
    "numpy>=1.26.0",
    
    # Text Processing
    "spacy>=3.7.0",
    "nltk>=3.8.0",
    "textstat>=0.7.0",
    
    # Data Validation
    "pydantic>=2.5.0",
    "marshmallow>=3.20.0",
    
    # PDF Generation
    "reportlab>=4.0.0",
    "weasyprint>=60.0",
    
    # Monitoring & Logging
    "sentry-sdk[django]>=1.39.0",
    "structlog>=24.1.0",
    "django-prometheus>=2.3.0",
    
    # Environment
    "python-dotenv>=1.0.0",
    "dj-database-url>=2.1.0",
    
    # Utilities
    "uuid7>=0.1.0",
    "python-dateutil>=2.8.0",
    "requests>=2.31.0",
    "httpx>=0.26.0",
    "pillow>=10.2.0",
    "whitenoise>=6.6.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-django>=4.7.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "factory-boy>=3.3.0",
    "freezegun>=1.4.0",
    "responses>=0.24.0",
    "faker>=23.0.0",
    
    # Linting & Formatting
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "django-stubs>=4.0.0",
    "djangorestframework-stubs>=3.14.0",
    
    # Debugging
    "django-debug-toolbar>=4.3.0",
    "ipdb>=0.13.0",
    "rich>=13.7.0",
    
    # Pre-commit
    "pre-commit>=3.6.0",
]

ai-local = [
    # Local AI processing
    "whisperx>=3.0.0",
    "torch>=2.2.0",
    "torchaudio>=2.2.0",
]

[project.scripts]
oracy-manage = "manage:main"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-django>=4.7.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "factory-boy>=3.3.0",
    "freezegun>=1.4.0",
    "responses>=0.24.0",
    "faker>=23.0.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "django-stubs>=4.0.0",
    "djangorestframework-stubs>=3.14.0",
    "django-debug-toolbar>=4.3.0",
    "ipdb>=0.13.0",
    "rich>=13.7.0",
    "pre-commit>=3.6.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "F",   # Pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "W",   # pycodestyle warnings
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "config.settings.development"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["."]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/venv/*",
    "*/.venv/*",
    "config/*",
    "manage.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
